backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
dnspython==2.6.1
python-nmap==0.7.1
pyOpenSSL==24.1.0
cryptography==42.0.5
requests==2.31.0
reportlab==4.1.0
sqlalchemy==2.0.29
aiosqlite==0.20.0
python-multipart==0.0.9
groq==0.9.0
qrcode[pil]==7.4.2
Pillow==10.3.0
